+++ src/utils/index.ts
@@ -3,10 +3,12 @@
 import * as vscode from 'vscode';
 import * as fs from 'fs';
 import { exec } from 'child_process';
-import { GNOME, GNOME_CLASSIC, KDE_PLASMA, platforms } from '../constants';
+import { GNOME, GNOME_CLASSIC, KDE_PLASMA, platforms, dictionary } from '../constants';
 
 const getSettings = (name: string): string|void => {
-  const settings = vscode.workspace.getConfiguration('open-native-terminal');
+  const settings = vscode.workspace.getConfiguration(dictionary.EXTENSION_NAME);
+  if (!settings) return;
+
   return settings.get(name);
 };
 
@@ -18,36 +20,34 @@ export const compose = (...fns) =>
 export const checkExistingPath = (path: string): string|void => {
   if (fs.existsSync(path)) {
     return path;
-  } else {
-    vscode.window.showErrorMessage('Oops, the path is wrong, please try again');
-    return;
   }
+
+  vscode.window.showErrorMessage(dictionary.WRONG_PATH);
 };
 
 export const checkEmptyPath = (path: string): string|void => {
   if (path) {
     return path;
-  } else {
-    vscode.window.showErrorMessage('Oops, the path is empty, please try again');
-    return;
   }
+
+  vscode.window.showErrorMessage(dictionary.EMPTY_PATH);
 };
 
 export const getCorrectPath = (path: string): string => {
   if (fs.lstatSync(path).isDirectory()) {
     return path.replace(/\s/g, '\\ ');
-  } else {
-    const pathToParentDir = path.replace(/(\/|\\)?([^\/\\]*)(\/*|\*)$/, '');
-    if (fs.lstatSync(pathToParentDir).isDirectory()) {
-      return pathToParentDir.replace(/\s/g, '\\ ');
-    } else {
-      getCorrectPath(pathToParentDir);
-    }
   }
+
+  const pathToParentDir = path.replace(/(\/|\\)?([^\/\\]*)(\/*|\*)$/, '');
+  if (fs.lstatSync(pathToParentDir).isDirectory()) {
+    return pathToParentDir.replace(/\s/g, '\\ ');
+  }
+
+  getCorrectPath(pathToParentDir);
 };
 
-export const chooseLinuxTerminal = (path: string): void => {
-  const defaultTerminal = getSettings('use-default-terminal');
+export const selectLinuxTerminal = (path: string): void => {
+  const defaultTerminal = getSettings(dictionary.SETTINGS_NAME);
   if (defaultTerminal) {
     exec(`cd ${path} && ${defaultTerminal}`);
     return;
@@ -66,14 +66,14 @@ export const chooseLinuxTerminal = (path: string): void => {
   }
 };
 
-export const checkCurrentOS = (path: string): void => {
+export const detectOSAndLaunchTerminal = (path: string): void => {
   switch(process.platform) {
     case platforms.mac:
       exec(`open -a Terminal.app "${path}"`);
     case platforms.win:
       exec(`start /D ${path} "%cd%" cmd`);
     default:
-      chooseLinuxTerminal(path);
+      selectLinuxTerminal(path);
   }
 };
 
@@ -83,4 +83,4 @@ export const getRootPath = (path: string): string => {
   }
 
   return getCorrectPath(path);
-}
+};